Basic code for autocompletion implementation in user input is finished. Added history and commands attributes to the main Chatbot class
